requirements.txt
psgtray
PySimpleGUI
python-dotenv
tendo
pyinstaller
